MachineLearning/lab3/requirements.txt
datasets==3.1.0
huggingface-hub==0.24.3
matplotlib==3.8.4
numpy==2.0.1
pandas==2.2.2
pillow==10.4.0
PyYAML==6.0.1
tqdm==4.66.4
safetensors==0.4.5
scikit-learn==1.5.1
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
diffusers==0.31.0